hotel/hotelapp/serializers.py
from django.conf.urls import url, include
from .models import Client, Room
from rest_framework import routers, serializers, viewsets


class ClientSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Client
        fields = '__all__'


class RoomSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Room
        fields = '__all__'
