hotel/requirements.txt
asgiref==3.2.3
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.3
django-cleanup==4.0.0
django-debug-toolbar==2.2
django-filter==2.2.0
django-phone-field==1.8.0
djangorestframework==3.11.0
Faker==0.9.1
idna==2.9
Markdown==3.2.1
mixer==6.1.3
Pillow==7.0.0
py-moneyed==0.8.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.8
